dockerflow-dashboard-backend/requirements.txt
fastapi==0.114.0
uvicorn[standard]==0.24.0
SQLAlchemy==2.0.20
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
python-multipart
